Bug #1522039 power: supply: Fix alerts for code scans
[root cause  ] Fix alerts for code scans
[changes     ]  charger-manager.c
[side effects] no
[self test   ]: pass
[download normally]:yes
[power on/off normally]:yes
[do common repository/branch inspection]:yes
[is there dependence]:no
[confirm dependent commit]:no
[board]: sp7731e_1h10_go2g
[test case]:poweron test
[reviewers   ] jinfeng/yongzhi
[change_type ] bringup --> sprd_chip_support
[tag_product ] common

+++ drivers/power/supply/charger-manager.c
@@ -5469,7 +5469,7 @@ static void cm_track_capacity_monitor(struct charger_manager *cm)
 			 */
 			capacity = (clbcnt - cm->track.start_clbcnt) / 1000;
 			capacity =
-				(total_cap * cm->track.start_cap) / 1000 + (u32)capacity;
+				(total_cap * (u32)cm->track.start_cap) / 1000 + (u32)capacity;
 
 			if (abs(capacity - total_cap) < total_cap / 2) {
 				set_batt_total_cap(cm, capacity);
